+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 ------------------
 
 - #2868 Fix auto CC contacts set in the client for new samples
+- #2867 Remove stale AT reference objects
+- #2866 Add transaction savepoints to large upgradesteps
+- #2865 Skip problematic worksheets in upgrade step 2711
 - #2862 Updated Webpack Dependencies
 - #2864 Fix upgrade step AT type removal calls
 - #2861 Allow to override results in auto-import